Fix other warnings.
When using -Wextra, there were some signedness related warnings, as well
as size warnings, like using %li for pointer diff on 32-bit systems.

This was build-tested on both a 32-bit ARM and 64-bit x86_64.

+++ rnet_message.h
@@ -23,8 +23,8 @@
 #include <stdint.h>
 
 struct rnet_message {
-	int len;
-	int alen;
+	size_t len;
+	size_t alen;
 	char buffer[];
 };
